requirements.txt
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
django-qrcode==0.3
future==0.18.1
gunicorn==19.9.0
idna==2.8
numpy==1.17.3
opencv-python-headless==4.1.1.26
Pillow==5.1.0
pycparser==2.19
python-telegram-bot==12.2.0
pyzbar==0.1.8
qrtools==0.0.2
requests==2.22.0
six==1.12.0
tornado==6.0.3
urllib3==1.25.6
